Add memory manager template parameter to allocator
As we add more memory managers to the code base, it will be useful to be
able to switch them out at compile time using template parameters. This
commit adds a new template parameter (and a sensible default) to the
allocator class, which can be used to configure the memory manager used
at compile time.