Made the C++ standard selection for the HIP compilation more stable.
Previously the C++ standard flag would go missing when reconfiguring
the build of the project.

It turned out that the CMAKE_INCLUDE_FLAG_SEP_HIP variable was only
causing problems at this stage. As already with CMake 3.18 the issue
was there, I was just blind to it in the Athena builds... But the
good news is that not setting that variable at all makes the
build setup correct. ;-)