Re-wrote the SYCL code compilation in the same way in which HIP is used.
This excludes us using triSYCL unfortunately, but I was never using
that thing for anything serious anyway. The Codeplay compiler on the
other hand could be taught to this type of configuration.

Since RPATH handling is now an issue for both the SYCL and HIP
binaries, for now updated vecmem_add_test(...) to explicitly set
LD_LIBRARY_PATH. Locking the project to Linux platforms for now.
Though implicitly it was already locked in...