Add basic definitions for memory resource setup
This commit lays the groundwork for the rest of the migration to a
memory resource architecture. That is to say, it defines some of the
common definitions and infrastructure on which we will build the memory
resources for CPU code, CUDA, and other paradigms. More specifically it
imports the required std::pmr types in a compiler agnostic way, and it
defines our custom abstract subclass of std::pmr::memory_resource. It
also defines a useful alias for creating vectors.