Link the CUDA library against the CUDA runtime
Previously, we relied on compiling most of the CUDA library with nvcc in
order to properly link it and include the headers. However, this
approach is not really compatible with the C++14 and C++17 dichotomy we
have in our code. Therefore, we have decided to compile as much as
possible using the standard C++ compiler. This requires us to link it
against the CUDA runtime target.