Make CUDA error handling private to the library
The CUDA error handling is the only header we have right now that
depends on the CUDA runtime, and having it as a public library makes it
so that external code relying on it needs to link against that runtime
as well. By making the error handling internal we can avoid this issue.