Separate CUDA container test kernel from driver
As it is right now, the CUDA container test contains both device code
and host driver code in the same translation unit, which forces us to
compile both with nvcc. This is normally not a problem, but as discussed
in #14 we need to compile our C++ code with the C++17 standard while we
can only compile our CUDA code with the C++14 standard. It is therefore
necessary to enforce a separation between these two contexts, and thus
we split it into two translation units. The C++ TU will handle the
driving of the kernel and interact with the C++17-based architecture of
the vecmem library, while the kernel itself will be compiled with C++14
and be agnostic of whatever C++ standard is being used outside.