Make Github Actions workflows compatible with act
In order to test the code locally according to the rules defined in the
Github Actions workflows, I am using act (not to be confused with ACTS,
see https://github.com/nektos/act) to simulate the workflows. However,
it seems that not all Github Actions code works properly with act. There
are some problems with build path substitutions that act doesn't seem to
accept. Some minor changes can make the workflows work with both act and
Github actions, which will make development of HIP and SYCL code a lot
easier for me.