Add an RAII CUDA device switching mechanism
This new class allows us to switch CUDA devices in an elegant, modern
C++ fashion that is scope-safe. When an instance of this class is
constructed, the CUDA device is set to whatever was passed as a
parameter, and the previously selected CUDA device is restored when the
object is destroyed.