Implement a contiguous downstream memory resource
There are a lot of use cases in which you might want to allocate
contiguous blocks of memory, but this is not supported by default by
many allocators. This new downstream memory resource guarantees that
consecutive allocations will be exactly contiguous in memory, which
makes it easier to allocate memory with efficient data copies and
locality in mind.