Added vecmem::sycl::device_memory_resource.
In order to make its interface not depend on SYCL types directly,
also introduced the vecmem::sycl::queue_wrapper and
vecmem::sycl::details::opaque_queue types.

Hid the vecmem::sycl::device_selector type from the public interface
of vecmem::sycl, and updated the affected source files to pick up its
(now private) header with a relative path for now. The tests will be
updated later on not to depend on this type directly.