Re-wrote vecmem::const_device_vector.
All of its features can now be used in device code (thanks to
vecmem::reverse_iterator), and added some more functions from
std::vector that could come in handy later on.

The object is now initialised using vecmem::const_device_vector_data.
The goal here is to "formalise" the communication of memory
payload between the host and device container objects. As this
design shall be very useful for multi-dimensional containers.