Add support for a resource-aware allocator
Currently, we support only allocation using memory resources via vectors
and other such containers. Sometimes this is too high-level, and we (or
users) may want to allocate objects individually using a memory
resource. This commit adds a new allocator class which, when initialized
with a memory resource, allows allocation of bytes directly, size for
object arrays, or constructs new objects.