Now (potentially) downloading GoogleTest during CMake configuration.
Switched from using ExternalProject_Add(...) for the GoogleTest build
to using FetchContent. With this, the code can decide during the CMake
configuration whether GTest is available on the system already, or
whether it needs to build GoogleTest as part of the project. All while
using GTest's CMake configuration.

Note that I still needed to declare a pair of ALIAS libraries to make
convenient use of the GTest libraries. But at least I didn't need to
declare imported libraries from scratch.