Require CMake 3.17 for the build of the CUDA library.
The FindCUDAToolkit.cmake module only became available in CMake 3.17.
Since I don't want to make the code more complicated than it needs to
be (by using FindCUDA.cmake with older CMake versions), it was easiest
to just increase the minimum CMake version when building the CUDA
library.

Note that if CUDA is not available, an older version of CMake will
also be able to handle the rest of the build.