Introduced vecmem::details::is_same_nc.
This trait can be used to detect two types that are the same, except
for their const-ness. This pattern is needed in the code in a few
places, so instead of implementing it in a complicated way in every
single place, it made more sense to introduce a simple custom trait
struct for it.

With the new trait in place, taught vecmem::device_array and
vecmem::device_vector how to construct a "constant version" of
themselves on top of a non-const data object.