Made VECMEM_DEBUG_MSG work without printing any variable values.
Previously the macro would always need to receive at least one
argument after the string literal that describes the format of
the printout.

I tried to use the ##__VA_ARGS__ GNU extension at first, but that
does not work with all compilers under all circumstances. This
trick on the other hand should...