Made the CMake code compatible with ROCm/HIP 4.1.
In ROCm/HIP 4.1 the previous "hcc" and "nvcc" values of the HIP_PLATFORM
environment variable for the hipcc compiler were replaced with "amd"
and "nvidia" respectively. With no backwards compatibility for "hcc"
or "nvcc". :-(

Seemingly this was the only change made. o.O The macros in the HIP
headers still use the HCC and NVCC naming internally.

For now I kept the default platform on "hcc". So HIP 4.1 users will
have to actively set it to something else.