Re-wrote the explicit CUDA memory copy code.
Most of the logic is now implemented as part of a vecmem::copy
class, which relies on a virtual function for performing low
level memory copies. This virtual function is overridden in
vecmem::cuda::copy to implement CUDA memory copies.

Updated the CUDA tests to the new memory copy UI.