Prevented the VecMem compiler flags to "bleed out" from the project.
Modified vecmem_add_flag(...) not to modify flags "globally", but
only in a given context. So that when we build VecMem as part of
another project, it wouldn't pollute the other project's build
options with its own. This required explicit include of
vecmem-compiler-options-cpp.cmake in a number of additional places.

At the same time I made some additional changes as well:
  - Updated the project's "build version" to 0.2.0, to be in sync with
    the previously created v0.1.0 git tag;
  - With the "bleeding" now stopped, I could remove the very specific
    C++ build flag that I had to add earlier to be able to compile
    GTest with the -pedantic flag;
  - Explicitly turned off the build (and the configuration) of the
    GoogleTest components that VecMem doesn't need. Including the
    installation of the GoogleTest libraries.