Added builds with all debug messages enabled.
So that formatting mistakes in the debug message declarations
would be detected by the CI. Plus to check that the results of
the unit tests do not change because of turning on the debug
messages.

At the same time renamed the "cuda_hip" build to "cuda-and-hip".
I just didn't like the "cuda_hip" name all too much...