Created tests for vecmem::jagged_device_vector with SYCL.
Mostly copying the CUDA test. But having to adapt it a bit to be able
to use it with the SYCL host device as well.

Unfortunately the tests do not work with an OpenCL iGPU device.
And OpenCL CPU devices didn't work before already. So I also updated
the CI configuration to only allow running the SYCL unit tests on
a Level-0 GPU device, or on the host directly.