Removed the size checks from the vector_view constructor.
One may legitimately want to create vecmem::data::vector_view
objects on top of non-host-accessible memory, in which case one
can not de-reference the m_size pointer.

And the code even uses the affected constructor for creating
non-resizable views in some situations. So checking for m_size
to not be a nullptr would also not be meaningful here.