Making the CMake SYCL build configuration a bit more flexible/robust.
Instead of hardcoding the usage of -fsycl and -fsycl-targets=<foo> into the
CMake files, the code now will only attempt to use -fsycl if no other instruction
is given to it using the SYCLCXX environment variable.

In order to control which executable to use for the build of the SYCL source
files, from now on the SYCLCXX environment variable should be used. Similar
to how CXX and CUDACXX can be.