Fix inclusion of PMR headers on C++14
It appears that the logic we use to determine whether we can import the
"real" or the experimental version of the standard PMR library is a
little broken. It does not seem to work on C++14, probably because
__has_include is a C++17 feature. However, I am not one hundred percent
sure what the exact problem is here. Adding these additional checks on
the C++ version seems to resolve the problem.

This commit fixes the issues outlined in acts-project/traccc#78.