Not forcing the usage of shared libraries in all cases anymore.
The code was updated not to necessarily build shared libraries in all
situations for the "main libraries". Since on Windows they really only
work as static libraries. At least at the moment.

On all other platforms, they are still set up as shared/dynamic libraries
by default. But the user is now allowed to override this behaviour from
the outside if necessary.