Tweaked the setup of the project's tests.
These changes are meant to make it easier to integrate VecMem into the build
of downstream projects. Allowing those projects to easily switch the VecMem
unit tests on or off, and to explicitly tell the VecMem build where it should
take GoogleTest from. (In case the tests *are* turned on.)