Switched to setting symbol visibility specifically for the main libraries.
This way VecMem would not affect the behaviour of client projects that
want to use it. Plus this is the only way to make the SYCL visibility
settings work. (For a variable like CMAKE_SYCL_VISIBILITY_PRESET to work,
the CMake C++ source code would need to know about this language.)