Upgraded the SYCL Windows configuration to use dpcpp-cl instead of dpcpp.
As I had to learn, while the dpcpp executable exists on Windows, the
recommended compiler to use is actually dpcpp-cl, which provides a
user interface much closer to that of Visual Studio's cl executable.

At the same time updated the configuration to make the DLL and executable
building with dpcpp-cl much more successful than dpcpp was ever able to
produce.