Made it possible to build vecmem_core into a DLL.
As the biggest change I had to outsource most of
vecmem::binary_page_memory_resource's code into an implementation
struct. This was necessary because MSVC seems to have a bug with having
an std::vector<std::unique_ptr<T> > type in an exported class. Instead
of trying to finess around that issue, I decided to move all of the
memory resource's implementation into
vecmem::details::binary_page_memory_resource_impl, which the exported
class would use behind the scenes. Hiding all of the implementation of
the class from its users.