Explicit Symbol Export, main branch (2021.09.30.) (#117)
* Set up the generation of "export headers" for all VecMem libraries.

At the same time turning the default C++ visibility to "hidden".

* Declared all classes/functions that need to be exported from vecmem::core.

* Made vecmem_cuda work with symbol hiding turned on.

This necessitated building cuda_error_handling.cpp as part of the
test executable (as well), since I really didn't want to make that
code part of the public interface of the vecmem_cuda library.

* Made vecmem_hip work with symbol hiding turned on.

This necessitated building hip_error_handling.cpp as part of the
test executable (as well), since I really didn't want to make that
code part of the public interface of the vecmem_hip library.

* Fixed the formatting of the source files.

* Set up additional variables controlling the SYCL compiler's behaviour.

Most importantly the variable responsible for setting the default
visibility of symbols implemented in SYCL source files.

* Switched to setting symbol visibility specifically for the main libraries.

This way VecMem would not affect the behaviour of client projects that
want to use it. Plus this is the only way to make the SYCL visibility
settings work. (For a variable like CMAKE_SYCL_VISIBILITY_PRESET to work,
the CMake C++ source code would need to know about this language.)

* Made the SYCL code work after turning on symbol hiding.

* Made the new SYCL language configuration work on Windows.

With CMake's own modules in use for setting up the properties of
DPC++ on Windows, the code had to be tweaked a bit. Including a
change in debug.hpp, as the SYCL build now gained the MSVC specific
definitions as a side-effect.