Memory Resource Updates, main branch (2021.10.04.) (#119)
* Introduced the vecmem::details::memory_resource_base type.

It is meant to serve as the base class of all of the concrete
memory resource classes.

* Updated the core memory resources to use vecmem::details::memory_resource_base.

At the same time added some additional documentation and debug
messages to some of them. (Not for vecmem::binary_page_memory_resource,
since for that I have other larger changes in the pipeline already.)

* Updated the CUDA memory resources to use vecmem::details::memory_resource_base.

At the same time added some additional documentation and debug
messages to all of them.

* Made some more memory resources "final".

* Updated the HIP memory resources to use vecmem::details::memory_resource_base.

At the same time added some debug messages to them, and tried to harmonise them
with the implementation of the other memory resources.

* Updated the SYCL memory resources to use vecmem::details::memory_resource_base.

At the same time tried to harmonise the behaviour/implementation of these
memory resources with all the other ones.