DLL Building Support, main branch (2021.10.03.) (#118)
* Upgraded the SYCL Windows configuration to use dpcpp-cl instead of dpcpp.

As I had to learn, while the dpcpp executable exists on Windows, the
recommended compiler to use is actually dpcpp-cl, which provides a
user interface much closer to that of Visual Studio's cl executable.

At the same time updated the configuration to make the DLL and executable
building with dpcpp-cl much more successful than dpcpp was ever able to
produce.

* Switched to building shared libraries on all platforms by default.

At the same time moving the test binaries of the project to the default
CMAKE_RUNTIME_OUTPUT_DIRECTORY, to make sure that the test executables
would easily find the possible DLLs produced by the build.

* Made it possible to build vecmem_core into a DLL.

As the biggest change I had to outsource most of
vecmem::binary_page_memory_resource's code into an implementation
struct. This was necessary because MSVC seems to have a bug with having
an std::vector<std::unique_ptr<T> > type in an exported class. Instead
of trying to finess around that issue, I decided to move all of the
memory resource's implementation into
vecmem::details::binary_page_memory_resource_impl, which the exported
class would use behind the scenes. Hiding all of the implementation of
the class from its users.

* Made it possible to build vecmem_cuda into a DLL.

It required disabling the warning about having exported classes
inherit from (non-exported) standard library types.

* Made it possible to build vecmem_sycl into a DLL.

It required disabling 2 warnings from MSVC about using
standard library types in exported classes.

* Adapted the MSVC warning manupulations to the updated code.

* Fixed the SYCL build on Linux after the Windows improvements.

At the same time removed a variable that the SYCL code inherited
from the HIP code, but which was never actually used in the
SYCL build.