Implement an instrumenting memory resource
This memory resource is a little different from the others, in the sense
that its purpose is not to control how memory is allocated, but rather
to monitor and instrument that process. This memory resource is mainly
designed for debugging, profiling, and testing other resources, both in
our code and in user code.

The instrumenting memory resource provides support for arbitrary
user-provided pre-allocation, post-allocation, and pre-deallocation
hooks. It keeps track of a chronological series of allocation and
deallocation requests, and it knows about the side of its own
outstanding allocation pool.