Add a special memory resource for debugging
This commit adds a new downstream memory resource which is designed to
make it easier to debug memory errors when using memory resources. The
debug memory resource can be wrapped around any other memory resource,
and will report (through the use of exceptions) on three types of
errors:

1. Overlapping allocation errors: if an allocation is returned upstream
   which overlaps with an existing allocation, an error is raised.
2. Invalid deallocation errors: if a deallocation request is issued
   which matches an outstanding allocation, but the allocation
   parameters (e.g. the size and alignment) are not the same, an error
   is raised.
3. Double deallocation errors: if a deallocation request is issued for
   memory which is not an outstanding allocation, an error is raised.

Please note that this class uses `std::logic_error`, and not
`std::bad_alloc`. This is on purpose, in order to allow us to
distinguish between different kinds of errors, and also to prevent the
exception from being swallowed by, for example, a coalescing memory
resource.