Made MSVC always set __cplusplus when using vecmem::core
Instead of setting the appropriate flag just for this project,
enabled the MSVC flag for turning on the setting of the
__cplusplus macro for all clients of vecmem::core. Including
clients outside of this project.

Also enabled the appropriate flag for CUDA / nvcc, so that on
Windows that would also always set __cplusplus when linking against
vecmem::core.

Added a test for all of this, by making use of the CUDA memory
resources using the NVCC compiler. (When that would be capable
of C++17 support.)

Finally, suppressed the warning about exporting a class that inherits
from a non-exported class, when using the NVCC compiler on Windows.