Implement a custom main function for CUDA tests
One of the goals of vecmem's tests is to make sure that there are no
memory leaks. For CPU code, we can test this by using valgrind, and for
CUDA we can test the same thing using `cuda-memcheck`. However, the CUDA
memchecker has a peculiar feature where the leak checker (`--leak-check
full`) only works if you make a call to `cudaDeviceReset` at the end of
your program. We currently rely on our tests being linked against the
GTest standard main function, which for obvious reasons does not include
this test. In this commit, I add a new CUDA-specific Google Test main
function which we can use instead, which makes the correct call to the
CUDA device reset function.