Add an implementation of unique pointers
We've had some code in vecmem that interfaces with std::unique_ptr for a
while now, but we don't really have a formal, outward-facing,
user-friendly implementation of unique pointers in vecmem. This commit
changes that, as it adds two different kinds of unique pointer to
vecmem. The first is a unique object pointer, which behaves very much
like std::unique_ptr, in that it handles the construction and
destruction of the relevant type. The second is a unique allocation
pointer, which works only for trivial types. The reaon the allocation
pointer type needs to exist is that it works for memory resources which
are not host-accessible.