Removed the custom constructors from vecmem::static_array.
Thereby making it possible to hold non-movable types (like vecmem::vector-s)
in such arrays. The setup of vecmem::static_array now very much resembles
the structure of std::array as it is implemented in libstdc++.

This meant that some incantations for constructing arrays no longer work,
requiring some small tweaks in the tests.

This also meant that the implementation of vecmem::static_vector had to
stop sharing the array type trait with vecmem::static_array.