Use new unique pointers internally
Previously, we were using some unique pointers internally for, for
example, our array types and our buffer types. This pull request changes
this to use the new `vecmem::unique_obj_ptr` and
`vecmem::unique_alloc_ptr` types, which are drop-in replacements. This
homogenizes the interal and external representation of these smart
pointers.