Manually copied @Yhatoh's update from #154 into a single commit.
Instead of trying to squash the commits from that PR into a single one,
and then rebase it on top of the current main branch, it seemed easier
to just create a new commit by hand.

These updates are meant to reduce the number of allocations from the
arena allocator. Also fixing an issue in the vecmem::details::arena
destructor, and hiding vecmem::details::minimum_superblock_size from
the outside world.