Taught the project how to use ComputeCpp for SYCL code compilation.
The project now determines a "vendor ID" for the SYCL compiler that was
set up. (Only knowing about "IntelLLVM" and "ComputeCpp" so far.)

A ComputeCpp specific configuration was created for Linux, which also
sets up the build to automatically link against ComputeCpp's runtime
library.

At the same time fixed an issue with how the C++ standard flags would
be given to the compilation. This was not working correctly for a
while. It's just that the Intel compiler uses C++17 by default...