SYCL Compatibility, main branch (2021.12.17.) (#158)
* Small tweaks to make the SYCL code "more compatible" with non-Intel compilers.

Hiding AMD OpenCL devices in vecmem::sycl::device_selector is useful for the
Intel compiler as well however.

* Taught the project how to use ComputeCpp for SYCL code compilation.

The project now determines a "vendor ID" for the SYCL compiler that was
set up. (Only knowing about "IntelLLVM" and "ComputeCpp" so far.)

A ComputeCpp specific configuration was created for Linux, which also
sets up the build to automatically link against ComputeCpp's runtime
library.

At the same time fixed an issue with how the C++ standard flags would
be given to the compilation. This was not working correctly for a
while. It's just that the Intel compiler uses C++17 by default...

* Taught vecmem::sycl::copy what to do if sycl::queue::memset is not available.

Which can be the case with certain SYCL compilers.

* Added "-x c++" to the compilation like of a "generic SYCL compiler".

Just to make it possible to achieve "some level of success" with hipSYCL,
without explicitly supporting it just yet.

* Made the SYCL compiler detection compatible with older CMake and oneAPI versions.