Implement an aligned multi-region allocator
This continues on the work done in #160, but extends the alignment
guarantees as well as the flexibility of the code. The new function
introduced is capable of allocating a block of memory containing a
series of regions, each of which has a type as well as a number of
elements, which all abide by the alignment rules of their respective
types.