Switched to only using -Werror for SYCL sources when it is safe to do so.
Under some circumstances experimental versions of Intel's compiler may be
printing warnings because of its own bugs. To allow Debug builds in such
configurations as well, if the build sees that even an empty .sycl file
produces a warning, then -Werror is not used.

To do this, the vecmem_check_sycl_code_compiles(...) function was made
just a little smarter.