Using sycl::atomic_ref for vecmem::atomic in case it is available.
When sycl::atomic_ref is available, vecmem::atomic is set up simply as
a thin layer above that class.

Since oneAPI only added support for sycl::atomic_ref relatively recently,
the CMake configuration needs to check whether that type is available
with the particular SYCL compiler that we use, or not.