Introduced vecmem::atomic_ref, and switched vecmem::device_vector to using it.
With modern SYCL the new type is simply an alias for sycl::atomic_ref, and
with C++20(+) it is a typedef of std::atomic_ref. For anything else it is
a carbon copy of vecmem::atomic, with a different constructor.

Updated the unit tests that explicitly make use of vecmem::atomic, to also
test vecmem::atomic_ref.