Add new overloads for unique alloc pointers
In order to expand the usability of vecmem's unique pointers, we want to
be able to use them without having to ensure that the type we're
allocating space for is trivially constructible. Rather, we want to
allow some situations in which trivial copyability is enough. This
commit does that: it adds two new overloads of `make_unique_alloc` which
allow the user to copy some existing memory to a device, so we don't
need to worry about constructing the objects in host-inaccessible
memory.