Add tests for new unique pointer copy semantics
We have new semantics for unique pointers which allow them to be used as
vehicles to copy memory from host to device. This commit adds a few new
tests to our test suite which ensures that this functionality works
correctly. It also tests these functions with non-trivially
constructible types, which we previously were unable to handle.