Add alignment support to contiguous resource
Previously, the contiguous memory resource ignored its allocation
parameter, which lead to errors - in particular issue #180. This commit
reworks the internal allocation strategy to use std::align, ensuring
that memory is properly aligned.

This also breaks (in a valid way) some of the tests, which are adapted
to support the new behaviour.

Closes #180.