Accelerate buddy allocator internals
As we have seen in #182, the buddy allocator we have can sometimes be
pretty slow. This was not due to an issue with the fundamental design of
the allocator, but rather with its internal structure and
implementation. This commit updates the allocator to use a contiguous
array of pages, which simplifies a lot of operations and reduces the
need for pointer chasing, and significantly improves the performance of
the allocator.

Relies on #185 and #186.