Switched to using std::aligned_alloc when it's available.
At the same time updated the documentation and debug messages
of the code a little.

Unfortunately std::aligned_alloc is not available with MSVC
(and on Windows in general from what I understand). So for that
case the code implements its own aligned allocation scheme,
which involves a bit of over-allocation.