Fix buddy allocator duplicate allocation bug
There was a bug with the binary page allocator where it would
incorrectly identify the correct page to deallocate, as it would always
dig down to the smallest possible page with the correct address rather
than finding the one with the correct size. This commit fixes the issue,
which was reported by Chamodya in #182.