Made copies from resizable jagged vectors possible.
The previously implemented code broke down if the source of
the copy was a resizable jagged_vector_buffer, which was not
filled with data completely.

In the updated setup the copies generally use the vector capacities
instead of the sizes, in order to have a chance for optimised
copies out of resizable jagged_vector_buffer-s.

At the same time introduced the vecmem::copy::set_sizes(...) function
for conveniently setting the sizes of a jagged vector (buffer).
Which can come in handy when it's the target container that is
resizable, and not the source one. (Though this latter operation
I didn't really test yet.)