Tweaked the output levels of the debugging messages.
Moved all messages about copies to level 1, memory allocations
and deallocations (except for host memory [de]allocations) to
level 2, and a bunch of other messages to various different levels.

Tweaked the code in some places along the way, as I came across
some things that I didn't like too much.

As a bonus, removed memory_manager.cpp, which was left in the
repository from the very earliest days of the project's development.