Only allow optimised jagged vector copies for fully contiguous buffers.
In order to reduce the chances of mistakenly trying to perform
a single memory copy over multiple memory allocations, the jagged
vector copy now only tries to be optimal if all of the elements of
the jagged vector are contiguous in memory. Since the chances of
this happening by chance are very-very low.

Updated the unit tests to explicitly test the optimised copies
using (just) host memory.

Also made a tiny improvement/change in the implementation of
vecmem::data::jagged_vector_buffer, making use of a standard algorithm
instead of using a hand-written for-loop.