Removed vecmem::sycl::device_selector.
The class was not compatible with the latest update of SYCL2020
anymore, and it was not worth it to update it to be compatible.
Since this class was not useful ever since the SYCL_DEVICE_FILTER
environment variable became a thing in oneAPI.

Updated all clients to account for this change.