Removed the CMake debug printouts for the native builds.
CMake 3.26 switched to a new way of storing the detailed output
of its tests. But as long as all of windows-lates, macos-latest and
ubuntu-latest don't switch to CMake 3.26, we can't switch to the
new formalism yet.

So for now it's easier to just disable this printout for the native
builds, and only keep them for the containerised ones.