Made the SYCL build work on Windows with CMake >=3.27.
Putting version numbers in the DLL names had to be disabled
completely, as the logic around DLL_NAME_WITH_SOVERSION seems
to be set on putting version numbers into the SYCL shared library
DLL names otherwise. :-/

While at it, I noticed that modern CMake versions no longer need
the hack of adding "/link" to the linking commands, so the code
now looks for a pre-existing "/link" or "-link" argument, and
only modifies the linking commands if necessary.