Copy / MR Improvements, main branch (2023.11.10.) (#247)
* Explicitly exporting the copy virtual functions.

With certain compilers this ends up being needed.

* Made vecmem::copy resize jagged vectors correctly.

Making sure that the inner vectors would use the same
memory resource that was assigned to the outer vector.

* Switch filter test to using a host memory resource.

On WSL2 asynchronous copying into shared memory (with an
NVIDIA backend) does not work for some reason. Leading to
a very weird crash.

* Synchronizing keywords on copy and resource classes.

Making sure that they all use the same set of keywords for
declaring their virtual functions, in all of the libraries.