Fixed constant iteration over a non-const jagged vector.
The code was not handling this case correctly before. Where
one starts from a non-const jagged vector type, but asks for
a constant iterator over that non-const container.

Added tests for iterating over vecmem::device_vector and
vecmem::jagged_device_vector, so such problems would not
occur in the future.