(Thrust) Pool Memory Resource, main branch (2023.12.27.) (#259)
* Introduced integer_math.hpp.

Moved some of the generic functions from the implementation of
binary_page_memory_resource into this new place. While adding
some additional functions, which would be needed for
pool_memory_resource.

* Introduced vecmem::pool_memory_resource.

It is a copy of thrust::mr::unsynchronized_pool_resource. With
some slight tweaks to integrate it into this project.

* Switched to the code from thrust::mr::disjoint_unsynchronized_pool_resource.

As this is the one that actually works with non-host-accessible
upstream memory resources. Beside being a bit simpler in its
implementation as well. (And without a bug that the unit tests
would identify!)

Moved vecmem::pool_options to vecmem::pool_memory_resource::options.

* Addressing Stephen's comments.

Made the integer math functions a little more robust,
and pointed to the Apache 2 license explicitly in the
documentation of vecmem::pool_memory_resource.