Skip some SoA tests on devices with no FP64 support.
The device tests with vecmem::testing::jagged_soa_container require
the device to be able to execute FP64 operations. (Since the SoA
container has a jagged vector of doubles.) Which is not the case for
absolutely every SYCL device that we test. So on those these tests
should just be skipped.