vecmem::abstract_event improvements.
Made the sycl_event and cuda_event types implicitly wait for their
underlying events in their destructors, if the user did not wait
for them explicitly. This is to avoid 99% of the asynchronous
errors that I encountered during debugging.

At the same time also made it possible to explicitly ignore such
events, for the rare case where it may be needed.

Finally, introduced VECMEM_FAIL_ON_ASYNC_ERRORS for building the
project in a mode where asynchronous errors (users not explicitly
ignoring or waiting for an event) would cause the program to
terminate.