Copy (Test) Improvements, main branch (2024.02.29.) (#270)
* Harmonized / merged the copy tests.

All of the "copy tests" are now instantiated in the same way,
inside of single compilation units.

* Improved the support for asynchronous SoA operations in vecmem::copy.

Unfortunately some issues still remain, which I'm in the process of
finding / fixing.

* vecmem::abstract_event improvements.

Made the sycl_event and cuda_event types implicitly wait for their
underlying events in their destructors, if the user did not wait
for them explicitly. This is to avoid 99% of the asynchronous
errors that I encountered during debugging.

At the same time also made it possible to explicitly ignore such
events, for the rare case where it may be needed.

Finally, introduced VECMEM_FAIL_ON_ASYNC_ERRORS for building the
project in a mode where asynchronous errors (users not explicitly
ignoring or waiting for an event) would cause the program to
terminate.

* Fixed remaining synhronization issues identified by the "event improvements".

* Disabling some test setups on Windows.

It seems that the implementation of CUDA managed memory is
a bit more fragile on Windows than on Linux.