Increase the robustness of `device_atomic_ref`
Currently, the functionality is `compare_exchange_strong` is broken
because it relies on the CUDA `atomicCAS` builtin which functions
fundamentally differently from the C++ STL version of the equivalent
code. Indeed, the C++ version returns true on a succesful swap and false
otherwise. The CUDA version always returns the old value. As such, if
the old value is false-like, e.g. 0, the `compare_exchange_strong`
function will always appear to fail, even if it succeeded. This commit
fixes the above issue. Also increases the robustness of other atomic
operations, adds new concepts, and adds new tests.